@@ -5,9 +5,7 @@ namespace App\Command;
 use App\Repository\TaxRuleRepository;
 use Symfony\Component\Console\Attribute\AsCommand;
 use Symfony\Component\Console\Command\Command;
-use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
-use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Style\SymfonyStyle;
 use Symfony\Contracts\Cache\CacheInterface;
@@ -20,9 +18,8 @@ class CacheWarmTaxCommand extends Command
 {
     public function __construct(
         private TaxRuleRepository $taxRuleRepository,
-        private CacheInterface $cache
-    )
-    {
+        private CacheInterface $cache,
+    ) {
         parent::__construct();
     }
 
