@@ -11,12 +11,13 @@ use Symfony\Contracts\Cache\CacheInterface;
 final class CacheController extends AbstractController
 {
     #[Route('/cache/tax', name: 'app_cache_tax')]
-    public function taxCache( TaxRuleRepository $taxRuleRepository, CacheInterface $cache): Response
+    public function taxCache(TaxRuleRepository $taxRuleRepository, CacheInterface $cache): Response
     {
         $cachedRules = $taxRuleRepository->FindCacheTaxRule($cache);
+
         return $this->render('cache/tax.html.twig', [
             'cached_rules' => $cachedRules,
-            'cache_key' => 'tax_rules_cache'
+            'cache_key' => 'tax_rules_cache',
         ]);
     }
 
@@ -25,16 +26,17 @@ final class CacheController extends AbstractController
     {
         $cache->delete('tax_rules_cache');
         $this->addFlash('success', 'Tax rule cache cleared');
+
         return $this->redirectToRoute('app_cache_tax');
     }
 
     #[Route('/cache/tax/warm', name: 'app_cache_tax_warm')]
-public function warmTaxCache(CacheInterface $cache, TaxRuleRepository $taxRuleRepository): Response
-{
-    $taxRuleRepository->findCacheTaxRule($cache);
-    
-    $this->addFlash('success', 'Cache warmed up');
-    
-    return $this->redirectToRoute('app_cache_tax');
-}
+    public function warmTaxCache(CacheInterface $cache, TaxRuleRepository $taxRuleRepository): Response
+    {
+        $taxRuleRepository->findCacheTaxRule($cache);
+
+        $this->addFlash('success', 'Cache warmed up');
+
+        return $this->redirectToRoute('app_cache_tax');
+    }
 }
