@@ -13,7 +13,7 @@ use Symfony\Contracts\Cache\CacheInterface;
 class ComplexCalcs extends AbstractController
 {
     #[Route('/complex/calc', name: 'complex_calc')]
-    public function index(TaxCalc $taxCalc, OvertimeCalc $overtimeCalc, ProportionCalc $proportionCalc, CacheInterface $cache): Response 
+    public function index(TaxCalc $taxCalc, OvertimeCalc $overtimeCalc, ProportionCalc $proportionCalc, CacheInterface $cache): Response
     {
         $periodId = 4;
         $tax = $taxCalc->getTax();
