@@ -17,52 +17,52 @@ class EmployeeType extends AbstractType
     public function buildForm(FormBuilderInterface $builder, array $options): void
     {
         $builder
-            ->add('fullName',TextType::class,[
+            ->add('fullName', TextType::class, [
                 'label' => 'Full name',
                 'constraints' => [
                     new Regex([
                         'pattern' => '/^[a-zA-Zа-яА-ЯёЁ\s\-]+$/u',
-                        'message' => 'Only letters, spaces and dashes'
-                    ])
-                    ]
+                        'message' => 'Only letters, spaces and dashes',
+                    ]),
+                ],
             ])
-            ->add('TIN', TextType::class,[
+            ->add('TIN', TextType::class, [
                 'label' => 'TIN',
-                'constraints'=>[
+                'constraints' => [
                     new Length([
-                        'min'=> 12,
-                        'max'=>12,
-                        'exactMessage'=> "TIN should contain exactly 12 numbers"
+                        'min' => 12,
+                        'max' => 12,
+                        'exactMessage' => 'TIN should contain exactly 12 numbers',
                     ]),
                     new Regex([
                         'pattern' => '/^\d+$/',
-                        'message' => 'Only numbers'
-                    ])
-                ] 
+                        'message' => 'Only numbers',
+                    ]),
+                ],
             ])
-            ->add('bankAccount',TextType::class,[
+            ->add('bankAccount', TextType::class, [
                 'label' => 'Bank Account',
-                'constraints'=>[
-                new Length([
-                    'min'=>20,
-                    'max'=>20,
-                    'exactMessage'=> 'Bank Account should contain exactly 20 numbers'
-                ]),
-                new Regex([
-                    'pattern' => '/^\d+$/',
-                    'message' => 'Only numbers'
-                ])
-                ]
+                'constraints' => [
+                    new Length([
+                        'min' => 20,
+                        'max' => 20,
+                        'exactMessage' => 'Bank Account should contain exactly 20 numbers',
+                    ]),
+                    new Regex([
+                        'pattern' => '/^\d+$/',
+                        'message' => 'Only numbers',
+                    ]),
+                ],
             ])
-            ->add('employmentType',ChoiceType::class,[
+            ->add('employmentType', ChoiceType::class, [
                 'label' => 'employment type',
-                'choices'=>[
+                'choices' => [
                     'Full time' => 'FULL_TIME',
                     'Part time' => 'PART_TIME',
                 ],
-                'constraints'=>[
-                    new NotBlank(['message'=>'Choose an option!'])
-                ]
+                'constraints' => [
+                    new NotBlank(['message' => 'Choose an option!']),
+                ],
             ])
         ;
     }
