@@ -15,31 +15,31 @@ class TaxRuleType extends AbstractType
     public function buildForm(FormBuilderInterface $builder, array $options): void
     {
         $builder
-            ->add('minAmount', NumberType::class,[
+            ->add('minAmount', NumberType::class, [
                 'constraints' => [
-                    new Range(['min' => 0,'minMessage'=>'Cannot be negative!'])
-                ]
+                    new Range(['min' => 0, 'minMessage' => 'Cannot be negative!']),
+                ],
             ])
-            ->add('maxAmount', NumberType::class,[
+            ->add('maxAmount', NumberType::class, [
                 'required' => false,
-                'attr' => ['placeholder' => 'Live blank for "..and higher"']
+                'attr' => ['placeholder' => 'Live blank for "..and higher"'],
             ])
-            ->add('rate', NumberType::class,[
+            ->add('rate', NumberType::class, [
                 'scale' => 2,
                 'constraints' => [
                     new Range([
                         'min' => 0,
                         'max' => 100,
-                        'notInRangeMessage' => 'Rate must be between 0 and 100!'
-                    ])
-                ]
+                        'notInRangeMessage' => 'Rate must be between 0 and 100!',
+                    ]),
+                ],
             ])
-            ->add('taxType', ChoiceType::class,[
+            ->add('taxType', ChoiceType::class, [
                 'choices' => [
                     'Income tax' => 'INCOME_TAX',
                     'Pension' => 'PENSION',
-                    'Social' => 'SOCIAL'
-                ]
+                    'Social' => 'SOCIAL',
+                ],
             ])
         ;
     }
